dio-azure-ecommerce/requirements.txt
streamlit
azure-storage-blob
python-dotenv
sqlalchemy
pyodbc
pandas
